==> Code/Advanced Neural Networks/itransformer+mamba/0_generate_df_A.ipynb <==
import pandas as pd

df = pd.read_csv(r'E:\Program Files (x86)\Desktop\论文\NIPS\All_data.csv')

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
pd.set_option('display.max_colwidth', None)
pd.set_option('display.width', 1000)

df.head(10)

df.shape

df.dtypes

df['Gender'] = df['Gender'].astype('float64')
mean_value = df['Gender'].mean()

df['Gender'] = df['Gender'].fillna(mean_value)

df['Age'] = df['Age'].astype('float64')
mean_value = df['Age'].mean()

df['Age'] = df['Age'].fillna(mean_value)

df['Height'] = df['Height'].astype('float64')
mean_value = df['Height'].mean()
df['Height'] = df['Height'].fillna(mean_value)

df['Weight'] = df['Weight'].astype('float64')
mean_value = df['Weight'].mean()
df['Weight'] = df['Weight'].fillna(mean_value)

df['BMI'] = df['BMI'].astype('float64')
mean_value = df['BMI'].mean()
df['BMI'] = df['BMI'].fillna(mean_value)

df.fillna(0, inplace=True)
if df.isnull().values.any():
    print("DataFrame 中含有 NaN 值")
else:
    print("DataFrame 中不含有 NaN 值")

# 将日期字符串转换为 datetime 类型
df['Medical Record Date'] = pd.to_datetime(df['Medical Record Date'], format='%Y/%m/%d')

# 转换 datetime 为时间戳（单位为秒），并转换为浮点数
df['Medical Record Date'] = df['Medical Record Date'].apply(lambda x: x.timestamp()).astype(float)

df.head(5)

df2 = df.drop(['Pain Relief Status','ERSO_Recom','IRSO_Recom','ERWO_Recom','IRWO_Recom','NSAIDs_Recom','A/A_Recom','Others_Recom'], axis=1)

normalized_df = (df2 - df2.mean()) / df2.std()

normalized_df['Pain Relief Status']=df['Pain Relief Status']
normalized_df['ERSO_Recom']=df['ERSO_Recom']
normalized_df['IRSO_Recom']=df['IRSO_Recom']
normalized_df['ERWO_Recom']=df['ERWO_Recom']
normalized_df['IRWO_Recom']=df['IRWO_Recom']
normalized_df['NSAIDs_Recom']=df['NSAIDs_Recom']
normalized_df['A/A_Recom']=df['A/A_Recom']
normalized_df['Others_Recom']=df['Others_Recom']

normalized_df.head(5)

normalized_df.shape

if normalized_df.isnull().values.any():
    print("DataFrame 中含有 NaN 值")
else:
    print("DataFrame 中不含有 NaN 值")

normalized_df.to_pickle('./df_A.pkl')

restored_df_A = pd.read_pickle('./df_A.pkl')

restored_df_A.shape

restored_df_A.head(5)
