==> Code/Advanced Neural Networks/itransformer+mamba/4_analysis_data.ipynb <==
import pandas as pd

df = pd.read_csv('./data/最终数据.csv')

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
pd.set_option('display.max_colwidth', None)
pd.set_option('display.width', 1000)

df.head(10)

df.shape

df.dtypes

df = df.dropna(subset=['疼痛缓解情况及用药后疼痛评分（1.完全缓解2.部分缓解3.轻度缓解4.无效）'])  # 删除分类标签缺失的行

# 将所有缺失值填充为0
df = df.fillna(0)

if df.isnull().values.any():
    print("DataFrame 中含有 NaN 值")
else:
    print("DataFrame 中不含有 NaN 值")

df.shape

df.columns

df['性别 '] = df['性别 '].astype('float64')

df['年龄'] = df['年龄'].astype('float64')

df['吸烟史（是1，否0）'] = df['吸烟史（是1，否0）'].astype('float64')

df['饮酒史（是1，否0）'] = df['饮酒史（是1，否0）'].astype('float64')

df['过敏史（是1，否）'] = df['过敏史（是1，否）'].astype('float64')

df['心血管风险（1是，0否）'] = df['心血管风险（1是，0否）'].astype('float64')

df['胃肠道风险（1是，0否）'] = df['胃肠道风险（1是，0否）'].astype('float64')

df['缓释强阿片'] = df['缓释强阿片'].astype('float64')
df['即释强阿片'] = df['即释强阿片'].astype('float64')
df['缓释弱阿片'] = df['缓释弱阿片'].astype('float64')
df['即释弱阿片'] = df['即释弱阿片'].astype('float64')
df['非甾体抗炎药'] = df['非甾体抗炎药'].astype('float64')

df['抗惊厥/抗抑郁类药物'] = df['抗惊厥/抗抑郁类药物'].astype('float64')
df['其他'] = df['其他'].astype('float64')
df['是否阿片耐受（是1，否0）'] = df['是否阿片耐受（是1，否0）'].astype('float64')
df['便秘（是1,否0）'] = df['便秘（是1,否0）'].astype('float64')
df['恶心或呕吐（是1,否0）'] = df['恶心或呕吐（是1,否0）'].astype('float64')
df['其他（是1,否0）'] = df['其他（是1,否0）'].astype('float64')
df['不良反应用药（是1,否0）'] = df['不良反应用药（是1,否0）'].astype('float64')
df['缓释强阿片.1'] = df['缓释强阿片.1'].astype('float64')
df['即释强阿片.1'] = df['即释强阿片.1'].astype('float64')
df['缓释弱阿片.1'] = df['缓释弱阿片.1'].astype('float64')
df['即释弱阿片.1'] = df['即释弱阿片.1'].astype('float64')
df['非甾体抗炎药.1'] = df['非甾体抗炎药.1'].astype('float64')
df['抗惊厥/抗抑郁类药物.1'] = df['抗惊厥/抗抑郁类药物.1'].astype('float64')
df['其他.1'] = df['其他.1'].astype('float64')

df['镇痛药物控制时长'] = df['镇痛药物控制时长'].astype('float64')

# 将日期字符串转换为 datetime 类型
df['入组日期'] = pd.to_datetime(df['入组日期'], format='%Y/%m/%d %H:%M')

# 转换 datetime 为时间戳（单位为秒），并转换为浮点数
df['入组日期'] = df['入组日期'].apply(lambda x: x.timestamp()).astype(float)

df.dtypes

df.groupby('姓名').size().max()

from sklearn.preprocessing import StandardScaler
float_cols = df.select_dtypes(include=['float64']).columns
float_cols = float_cols.drop('疼痛缓解情况及用药后疼痛评分（1.完全缓解2.部分缓解3.轻度缓解4.无效）')

scaler = StandardScaler()

# 对特定列进行标准化处理
df.loc[:, float_cols] = scaler.fit_transform(df[float_cols])

df['疼痛缓解情况及用药后疼痛评分（1.完全缓解2.部分缓解3.轻度缓解4.无效）'] -= 1 # 把标签转换为从0开始
df['疼痛缓解情况及用药后疼痛评分（1.完全缓解2.部分缓解3.轻度缓解4.无效）'] = df['疼痛缓解情况及用药后疼痛评分（1.完全缓解2.部分缓解3.轻度缓解4.无效）'].astype('int')

if df.isnull().values.any():
    print("DataFrame 中含有 NaN 值")
else:
    print("DataFrame 中不含有 NaN 值")

df.to_pickle('./df_class.pkl')

df = pd.read_csv('./data/最终数据.csv')

df.shape

# 将所有缺失值填充为0
df = df.fillna(0)

if df.isnull().values.any():
    print("DataFrame 中含有 NaN 值")
else:
    print("DataFrame 中不含有 NaN 值")

df['性别 '] = df['性别 '].astype('float64')
df['年龄'] = df['年龄'].astype('float64')
df['吸烟史（是1，否0）'] = df['吸烟史（是1，否0）'].astype('float64')
df['饮酒史（是1，否0）'] = df['饮酒史（是1，否0）'].astype('float64')
df['过敏史（是1，否）'] = df['过敏史（是1，否）'].astype('float64')
df['心血管风险（1是，0否）'] = df['心血管风险（1是，0否）'].astype('float64')
df['胃肠道风险（1是，0否）'] = df['胃肠道风险（1是，0否）'].astype('float64')
df['缓释强阿片'] = df['缓释强阿片'].astype('float64')
df['即释强阿片'] = df['即释强阿片'].astype('float64')
df['缓释弱阿片'] = df['缓释弱阿片'].astype('float64')
df['即释弱阿片'] = df['即释弱阿片'].astype('float64')
df['非甾体抗炎药'] = df['非甾体抗炎药'].astype('float64')

df['抗惊厥/抗抑郁类药物'] = df['抗惊厥/抗抑郁类药物'].astype('float64')
df['其他'] = df['其他'].astype('float64')
df['是否阿片耐受（是1，否0）'] = df['是否阿片耐受（是1，否0）'].astype('float64')
df['便秘（是1,否0）'] = df['便秘（是1,否0）'].astype('float64')
df['恶心或呕吐（是1,否0）'] = df['恶心或呕吐（是1,否0）'].astype('float64')
df['其他（是1,否0）'] = df['其他（是1,否0）'].astype('float64')
df['不良反应用药（是1,否0）'] = df['不良反应用药（是1,否0）'].astype('float64')
df['缓释强阿片.1'] = df['缓释强阿片.1'].astype('float64')
df['即释强阿片.1'] = df['即释强阿片.1'].astype('float64')
df['缓释弱阿片.1'] = df['缓释弱阿片.1'].astype('float64')
df['即释弱阿片.1'] = df['即释弱阿片.1'].astype('float64')
df['非甾体抗炎药.1'] = df['非甾体抗炎药.1'].astype('float64')
df['抗惊厥/抗抑郁类药物.1'] = df['抗惊厥/抗抑郁类药物.1'].astype('float64')
df['其他.1'] = df['其他.1'].astype('float64')

# 替换列'镇痛药物控制时长'中的空字符串和空白字符串为0
df['镇痛药物控制时长'] = df['镇痛药物控制时长'].replace(['', ' '], 0)
df['镇痛药物控制时长'] = df['镇痛药物控制时长'].astype('float64')


# 将日期字符串转换为 datetime 类型
df['入组日期'] = pd.to_datetime(df['入组日期'], format='%Y/%m/%d %H:%M')

# 转换 datetime 为时间戳（单位为秒），并转换为浮点数
df['入组日期'] = df['入组日期'].apply(lambda x: x.timestamp()).astype(float)

df.dtypes

df.groupby('姓名').size().max()

df_re_1 = df.copy()

from sklearn.preprocessing import StandardScaler
float_cols = df_re_1.select_dtypes(include=['float64']).columns
float_cols = float_cols.drop('缓释强阿片.1')
scaler = StandardScaler()

# 对特定列进行标准化处理
df_re_1.loc[:, float_cols] = scaler.fit_transform(df_re_1[float_cols])

df_re_1.to_pickle('./df_re_1.pkl')

df_re_2 = df.copy()

float_cols = df_re_2.select_dtypes(include=['float64']).columns
float_cols = float_cols.drop('即释强阿片.1')
scaler = StandardScaler()

# 对特定列进行标准化处理
df_re_2.loc[:, float_cols] = scaler.fit_transform(df_re_2[float_cols])
df_re_2.to_pickle('./df_re_2.pkl')

df_re_3 = df.copy()

float_cols = df_re_3.select_dtypes(include=['float64']).columns
float_cols = float_cols.drop('缓释弱阿片.1')
scaler = StandardScaler()

# 对特定列进行标准化处理
df_re_3.loc[:, float_cols] = scaler.fit_transform(df_re_3[float_cols])
df_re_3.to_pickle('./df_re_3.pkl')

df_re_4 = df.copy()

float_cols = df_re_4.select_dtypes(include=['float64']).columns
float_cols = float_cols.drop('即释弱阿片.1')
scaler = StandardScaler()

# 对特定列进行标准化处理
df_re_4.loc[:, float_cols] = scaler.fit_transform(df_re_4[float_cols])
df_re_4.to_pickle('./df_re_4.pkl')

df_re_5 = df.copy()

float_cols = df_re_5.select_dtypes(include=['float64']).columns
float_cols = float_cols.drop('非甾体抗炎药.1')
scaler = StandardScaler()

# 对特定列进行标准化处理
df_re_5.loc[:, float_cols] = scaler.fit_transform(df_re_5[float_cols])
df_re_5.to_pickle('./df_re_5.pkl')

df_re_6 = df.copy()

float_cols = df_re_6.select_dtypes(include=['float64']).columns
float_cols = float_cols.drop('抗惊厥/抗抑郁类药物.1')
scaler = StandardScaler()

# 对特定列进行标准化处理
df_re_6.loc[:, float_cols] = scaler.fit_transform(df_re_6[float_cols])
df_re_6.to_pickle('./df_re_6.pkl')

df_re_7 = df.copy()

float_cols = df_re_7.select_dtypes(include=['float64']).columns
float_cols = float_cols.drop('其他.1')
scaler = StandardScaler()

# 对特定列进行标准化处理
df_re_7.loc[:, float_cols] = scaler.fit_transform(df_re_7[float_cols])
df_re_7.to_pickle('./df_re_7.pkl')

restored_df_A = pd.read_pickle('./df_class.pkl')
restored_df_A.head(10)

restored_df_A = pd.read_pickle('./df_re_1.pkl')
restored_df_A.head(10)
